FIX dtype for idx in NN

+++ python/cuml/cuml/neighbors/nearest_neighbors.pyx
@@ -660,7 +660,7 @@ class NearestNeighbors(UniversalBase,
                 D_ndarr.index = index
                 I_ndarr = cuml.common.input_to_cuml_array(
                     I_cparr,
-                    convert_to_dtype=(np.float32 if convert_dtype
+                    convert_to_dtype=(np.int64 if convert_dtype
                                       else None),
                 ).array
                 I_ndarr.index = index
